Modified week 1

+++ week1/01A.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-  int a, b, c, d;
+  double a, b, c, d;
   cin >> a >> b >> c >> d;
 
   double rataan = (a + b + c + d) / 4.0;
